separate the scrypt stuff into a separate lib that can be tested separately
